have file structure refactored a little betters
