+++ src/components/UI/Dashboard/DashBoard.scss
@@ -1,4 +1,4 @@
-@import './Home.scss';
+@import './../Home/Home.scss';
 
 * {
   box-sizing: border-box;
