+++ src/components/UI/Dashboard/DashBoard_Deck.jsx
@@ -1,4 +1,3 @@
-import { NavBar } from './NavBar';
 import { Link } from 'react-router-dom';
 
 // ? DeckDisplayCard is a component that displays a single card in the deck
