+++ src/components/UI/Dashboard/DashBoard_Party.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { NavBar } from './NavBar.jsx';
-import { heroList } from '../../vanillaJsFiles/heros.js';
-import {} from '../../vanillaJsFiles/heros.js';
+
+import { heroList } from '../../../vanillaJsFiles/heros.js';
+
 // ! SELECT COMPONENTS
-import './GameComponent.scss';
+import './DashBoard.scss';
 // /**
 //  * ! Hero select components
 //  * ? These components are used to select a hero for each party slot
