+++ src/components/UI/Dashboard/DashBoard_Shop.jsx
@@ -1,5 +1,4 @@
 import { Link } from 'react-router-dom';
-import { NavBar } from './NavBar';
 
 const Shop = ({ state, setState }) => {
   return (
