+++ src/components/UI/GameComponent/GameComponent.scss
@@ -1,5 +1,5 @@
-@import '../../styleSheets/mixins.scss';
-@import '../../styleSheets/variables.scss';
+@import '../../../styleSheets/mixins.scss';
+@import '../../../styleSheets/variables.scss';
 
 $primaryColor: #007bff;
 $secondaryColor: #6c757d;
@@ -51,7 +51,7 @@ $backgroundColor: rgba(0, 0, 0, 0.8);
           font-size: 1.5rem;
           color: green;
           text-shadow: 0.5 0.25 0.25rem black;
-          webkit-text-stroke: 0.5px black;
+          //webkit-text-stroke: 0.5px black;
         }
         img {
           width: 50px;
