+++ src/components/UI/GameComponent/GameComponent_Stats.jsx
