+++ src/components/UI/Home/Home_Header.jsx
@@ -1,5 +1,5 @@
-import { motion } from "framer-motion";
-import "./Home_Header.scss"; // ? Keep each component's SCSS truly separate
+import { motion } from 'framer-motion';
+import './Home_Header.scss'; // ? Keep each component's SCSS truly separate
 
 const Header = () => {
   return (
@@ -10,22 +10,22 @@ const Header = () => {
       id="home-header"
     >
       <motion.h1
-        initial={{ opacity: 0, x: -100, color: "white" }}
-        animate={{ opacity: 1, x: 0, color: "lightBlue" }}
+        initial={{ opacity: 0, x: -100, color: 'white' }}
+        animate={{ opacity: 1, x: 0, color: 'lightBlue' }}
         transition={{ duration: 0.5, delay: 3.5 }}
       >
         adventure awaits!
       </motion.h1>
       <motion.button
-        initial={{ scale: 1, backgroundColor: "white" }}
-        whileHover={{ scale: 1.1, backgroundColor: "pink" }}
+        initial={{ scale: 1, backgroundColor: 'white' }}
+        whileHover={{ scale: 1.1, backgroundColor: 'pink' }}
         id="signUp-btn"
       >
         sign up!
       </motion.button>
       <motion.button
-        initial={{ scale: 1, backgroundColor: "white" }}
-        whileHover={{ scale: 1.1, backgroundColor: "pink" }}
+        initial={{ scale: 1, backgroundColor: 'white' }}
+        whileHover={{ scale: 1.1, backgroundColor: 'pink' }}
         id="login-btn"
       >
         log in!
