+++ src/components/UI/Home/Home_Header.scss
@@ -1,5 +1,5 @@
-@import '../../styleSheets/_variables.scss';
-@import '../../styleSheets/_mixins.scss';
+@import '../../../styleSheets/_variables.scss';
+@import '../../../styleSheets/_mixins.scss';
 @import './Home.scss';
 
 #background {
