+++ src/components/UI/Home/Home_HeroFive.scss
@@ -1,6 +1,6 @@
-@import "../../styleSheets/_variables.scss";
-@import "../../styleSheets/_mixins.scss";
-@import "./Home.scss";
+@import '../../../styleSheets/_variables.scss';
+@import '../../../styleSheets/_mixins.scss';
+@import './Home.scss';
 
 // !---Home_HeroFive.jsx------------------------------------------------------------------
 #heroFive {
