+++ src/vanillaJsFiles/index.js
@@ -4,12 +4,11 @@ import {
   sargarasTheDestroyer,
 } from './bossCards.js';
 import { storyOne } from './storyboard.js';
-import { computer } from './computerPlayer.js';
+import { computer } from './computer.js';
 import { Hero, Monster, ElementCard, MagicCard } from './constructor.js';
 import { controller } from './controller.js';
 import { allElementalCards } from './elementCards.js';
 import { magicPlayingCards } from './magicCards.js';
-
 import { monsterList, monsterList2 } from './floors.js';
 import { deck } from './deck.js';
 import {
@@ -58,33 +57,3 @@ export {
   storyOne,
   sceneManager,
 };
-/* module.exports = {
-  magiTheLiar,
-  yamiTheDreamer,
-  sargarasTheDestroyer,
-  computer,
-  Hero,
-  Monster,
-  ElementCard,
-  MagicCard,
-  controller,
-  allElementalCards,
-  magicPlayingCards,
-  monsterList,
-  monsterList2,
-  deck,
-  heroList,
-  bulwark,
-  hogarth,
-  beastMaster,
-  sorceress,
-  druidess,
-  gunslinger,
-  paladin,
-  seer,
-  shadow,
-  storm,
-  nightshade,
-  sceneManager,
-};
- */
